Fix missing coin values in statistics, sort positions by percentage, change coin quantity to Integer
- Coins linked to an asset with no depot holdings now appear as own positions instead of being silently dropped
- COIN positions included in by-index statistics view
- All statistics views and dashboard sorted by percentage descending
- Coin quantity changed from BigDecimal to Integer

+++ src/main/java/de/wsc/wealth/controller/DepotController.java
@@ -104,7 +104,7 @@ public class DepotController {
         BigDecimal coinTotal = BigDecimal.ZERO;
         for (Coin c : coins) {
             if (c.getMetal() == null || c.getWeightGrams() == null || c.getQuantity() == null) continue;
-            BigDecimal oz = c.getWeightOz().multiply(c.getQuantity());
+            BigDecimal oz = c.getWeightOz().multiply(BigDecimal.valueOf(c.getQuantity()));
             coinOzByMetal.merge(c.getMetal(), oz, BigDecimal::add);
             BigDecimal val = coinService.valueEur(c, spotPrices);
             if (val != null) {
